+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+##### Added 
+- Submission parameter now passed in offline-capable views too.
+
 [1.23.2] - 2016-02-10
 ----------------------
 ##### Changed